+++ matlab/utilities-1stparty/aircraftregistry/readfaaacreg.m
@@ -1,5 +1,5 @@
 function readfaaacreg(varargin)
-% Copyright 2018 - 2020, MIT Lincoln Laboratory
+% Copyright 2018 - 2021, MIT Lincoln Laboratory
 % SPDX-License-Identifier: BSD-2-Clause
 
 %% Input parser
@@ -16,7 +16,20 @@ parse(p,varargin{:});
 
 %% Load aircraft reference file
 fid = fopen([p.Results.inDir filesep 'ACFTREF.txt'],'r'); % Open File
-textACREF = textscan(fid,'%07.0s %030.0s %020.0s %01.0s %02.0s %01.0s %01.0s %02.0s %03.0f %07.0s %04.0s','HeaderLines',1,'Delimiter',',');
+
+% Based on number of characters in header row, read in data
+% The FAA changed the aircraft reference file in 2020, the switch / case is
+% to promotes backwards compatibility with data prior to 2020/12/03
+% https://github.com/Airspace-Encounter-Models/em-core/issues/4
+textHeader = fgetl(fid); % Get headerline
+switch numel(strfind(textHeader,','))
+    case 11
+        textACREF = textscan(fid,'%07.0s %030.0s %020.0s %01.0s %02.0s %01.0s %01.0s %02.0s %03.0f %07.0s %04.0s','HeaderLines',0,'Delimiter',',');
+    case 13
+        textACREF = textscan(fid,'%07.0s %030.0s %020.0s %01.0s %02.0s %01.0s %01.0s %02.0s %03.0f %07.0s %04.0s %015.0s %050.0s','HeaderLines',0,'Delimiter',',');
+    otherwise
+        error('readfaaacreg:acreflen','%i fields in aircraft reference file, was expecting 11 or 13 fields',numel(strfind(textHeader,',')));
+end
 fclose(fid); % Close file
 
 % Parse